# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anime Streaming API",
    "description": "API for streaming anime with multi-season support",
    "version": "18.2.1"
  },
  "paths": {
...
    "/local-overview/{token}": {
      "get": {
        "tags": [
          "local"
        ],
        "summary": "Get Local Overview",
        "description": "Metadata and episodes, or a play descriptor for a movie.",
        "operationId": "get_local_overview_local_overview__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/local-browse": {
      "get": {
        "tags": [
          "local"
        ],
        "summary": "Get Local Browse",
        "description": "The children of one directory, or the enabled roots. ``title`` entries get\nthe cached art; ``folder`` and ``file`` entries are the raw fallback for media\nthat never resolved to a title.",
        "operationId": "get_local_browse_local_browse_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Directory token; omit for the source-root level",
              "title": "Token"
            },
            "description": "Directory token; omit for the source-root level"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search/local": {
      "get": {
        "tags": [
          "local"
        ],
        "summary": "Search Local",
        "description": "Shaped like the other /search/* surfaces. Matches the enriched titles, so a\n``tmdb-<id>`` folder is findable by its real name once its art is cached.",
        "operationId": "search_local_search_local_get",
        "parameters": [
          {
            "name": "query_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Local title/filename to search",
              "title": "Query Name"
            },
            "description": "Local title/filename to search"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/watch-local/{token}": {
      "get": {
        "tags": [
          "local"
        ],
        "summary": "Watch Local",
        "description": "The /watch NDJSON contract for one file, so the player needs no second\npipeline: ``meta``, at most one ``stream``, then ``done``.",
        "operationId": "watch_local_watch_local__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Title"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}